Make the other classes use the renamed version.
